requirements.txt
selenium==4.12.0
vonage==2.13.0
python-dotenv==1.0.0